requirements.txt
asgiref==3.8.1
bleach==6.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.0.0
Django==4.0.4
django-cors-headers==3.13.0
django-filter==21.1
django-multiselectfield==0.1.12
django-rest-passwordreset==1.2.1
django-sortedm2m==3.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
drf_api_logger==1.1.10
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psycopg2==2.9.9
PyJWT==2.9.0
python-decouple==3.6
pytz==2024.1
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
webencodings==0.5.1
whitenoise==6.6.0
